+++ R/r_repos.R
@@ -3,11 +3,18 @@
 #' Report on which repositories R packages are distributed through
 #'  (i.e. base R, CRAN, Bioc, rOpenSci, R-Forge, and/or GitHub).
 #' @param which Which R repositories to extract data from.
+#' @param add_downloads Add the number of downloads from each repository.
+#' @param add_descriptions Add metadata from \emph{DESCRIPTION} files.
+#' @param add_github Add metadata from the respective GitHub repository 
+#' for each R package (if any exists).
 #' @param upset_plot Whether to create an upset plot 
 #' showing R package overlap between repositories.
 #' @param show_plot Print the plot.
 #' @param save_path Path to save upset plot to.
 #' @param verbose Print messages.
+#' @param height Saved plot height.
+#' @param width Saved plot width.
+#' @param nThread Number of threads to parallelise data queries across.
 #' @inheritParams BiocManager::repositories
 #' @returns Named list.
 #' 
@@ -17,9 +24,15 @@
 #' report <- r_repos()
 r_repos <- function(which=r_repos_opts(),
                     version=NULL,
+                    add_downloads=FALSE,
+                    add_descriptions=FALSE,
+                    add_github=FALSE,
                     upset_plot=TRUE,
                     show_plot=TRUE,
                     save_path=tempfile(fileext = "upsetr.pdf"),
+                    height=7,
+                    width=10,
+                    nThread=1,
                     verbose=TRUE){
     
     if(isTRUE(upset_plot)) requireNamespace("UpSetR")
@@ -28,7 +41,11 @@ r_repos <- function(which=r_repos_opts(),
     
     #### Gather data ####
     pkgs <- r_repos_data(which = which,
+                         add_downloads =  add_downloads,
+                         add_descriptions = add_descriptions,
+                         add_github = add_github,
                          version = version,
+                         nThread = nThread,
                          verbose = verbose)  
     #### Upset plot ####
     if(isTRUE(upset_plot)){
@@ -36,6 +53,8 @@ r_repos <- function(which=r_repos_opts(),
                                save_path = save_path, 
                                show_plot = show_plot,
                                verbose = verbose,
+                               height = height,
+                               width = width,
                                sets.bar.color = "slategrey",
                                main.bar.color = "slategrey",
                                text.scale = 1.5,