+++ R/r_repos_data.R
@@ -3,8 +3,6 @@
 #' Gather data on which repositories R packages are distributed through
 #'  (e.g. CRAN, Bioc, rOpenSci, and/or GitHub).
 #' @param include A subset of packages to return data for.
-#' @param add_downloads Add the number of downloads from each repository.
-#' @param add_descriptions Add metadata extracted from \emph{DESCRIPTION} files.
 #' @param cast Cast the results to wide format
 #'  so that each package only appears in one row.
 #' @param nThread Number of threads to parallelise \code{add_descriptions} 
@@ -18,104 +16,46 @@
 #' @importFrom utils installed.packages available.packages
 #' @importFrom data.table merge.data.table
 #' @examples 
-#' #### Without downloads data ####
-#' pkgs <- r_repos_data()
-#' 
-#' #### With downloads data ####
-#' include <- pkgs[r_repo=="Bioc",][seq_len(10),]$package
-#' downloads <- r_repos_data(include=include,
-#'                           add_downloads = TRUE)
+#' #### All packages ####
+#' pkgs1 <- r_repos_data()
+#' #### Specific packages ####
+#' #### Add extra data ####
+#' pkgs3 <- r_repos_data(include="echogithub",
+#'                       which = r_repos_opts(exclude = NULL),
+#'                       add_downloads = TRUE,
+#'                       add_descriptions = TRUE,
+#'                       add_github = TRUE,
+#'                       cast=TRUE)
 r_repos_data <- function(include=NULL,
                          add_downloads=FALSE,
                          add_descriptions=FALSE,
+                         add_github=FALSE,
                          which=r_repos_opts(),
                          cast=FALSE,
                          version=NULL,
                          nThread=1,
                          verbose=TRUE){
-    requireNamespace("rvest")
-    requireNamespace("BiocManager")
-    requireNamespace("githubinstall")
+    # templateR:::source_all()
+    # templateR:::args2vars(r_repos_data) 
     installed <- package <- NULL;
     
-    which <- tolower(which)
-    res <- list()
-    #### Base R ####
-    if("base" %in% which){
-        messager("Gathering R packages: base",v=verbose)
-        res[["base"]] <- data.table::data.table(
-            package=rownames(utils::installed.packages(priority="base"))
-        ) 
-    } 
-    #### CRAN ####
-    if("cran" %in% which){
-        messager("Gathering R packages: CRAN",v=verbose)
-        cran <- utils::available.packages(
-            contriburl = "https://cran.rstudio.com/src/contrib") |> data.frame()
-        res[["CRAN"]] <- data.table::data.table(package=cran$Package)
-    } 
-    #### Bioc ####
-    if("bioc" %in% which){
-        messager("Gathering R packages: Bioc",v=verbose)
-        # *Note*: This only retrieves Bioc packages in the currently installed 
-        # release of Bioconductor. Packages that are only in older versions of Bioc 
-        # (and were later deprecated) will not be listed here. 
-        ### This function gives all packages, including CRAN, Bioc,
-        # and anything currently installed. 
-        # bioc <- BiocManager::available()  
-        ### This only gives Bioc packages ####
-        if(is.null(version)) version <- BiocManager::version()
-        repos <- suppressMessages(
-            BiocManager::repositories(version = version)
-        )
-        repos <- repos[names(repos)!="CRAN"]
-        bioc <- utils::available.packages(repos = repos) |> data.frame()
-        res[["Bioc"]] <- data.table::data.table(package=rownames(bioc))
-    } 
-    #### rOpenSci ####
-    if("ropensci" %in% which){
-        messager("Gathering R packages: rOpenSci",v=verbose)
-        res[["rOpenSci"]] <- 
-            data.table::data.table(
-                package=rvest::read_html("https://docs.ropensci.org/") |>
-                    rvest::html_element("#repolist") |>
-                    rvest::html_children() |> 
-                    rvest::html_text() 
-                )
-    }  
-    #### R-forge #### 
-    if("rforge" %in% which){
-        messager("Gathering R packages: R-Forge",v=verbose)
-        rforge <- utils::available.packages(
-            repos = "http://R-Forge.R-project.org") |> data.frame()  
-        res[["R-Forge"]] <- data.table::data.table(package=rownames(rforge))
+    if(isTRUE(add_github) && 
+       isFALSE(add_descriptions)){
+        messager("add_descriptions must =TRUE when add_github=TRUE.",
+                 "Setting add_descriptions=TRUE.",
+                 v=verbose)
+        add_descriptions <- TRUE
     }
-    #### GitHub ####
-    if("github" %in% which){  
-        messager("Gathering R packages: GitHub",v=verbose)
-        # githubinstall::gh_update_package_list()
-        github <- githubinstall::gh_list_packages() 
-        res[["GitHub"]] <- data.table::data.table(package=github$package_name)
-    }
-    #### GitHub ####
-    if("local" %in% which){  
-        messager("Gathering R packages: local",v=verbose)
-        # githubinstall::gh_update_package_list()
-        local <- utils::installed.packages()
-        res[["local"]] <- data.table::data.table(package=rownames(local))
-    }
-    #### Merge all repos #### 
-    pkgs <- data.table::rbindlist(res, 
-                                  fill = TRUE,
-                                  use.names = TRUE, 
-                                  idcol = "r_repo") 
-    if(!is.null(include)) pkgs <- pkgs[package %in% include]
-    #### Add installed info ####
-    pkgs[,installed:=package %in% rownames(utils::installed.packages())]
+    #### List all R packages ####
+    pkgs <- r_repos_list(which = which, 
+                         include = include,
+                         version = version,
+                         verbose = verbose) 
     #### Add downloads ####
     if(isTRUE(add_downloads)){
         pkgs <- r_repos_downloads(pkgs = pkgs,
                                   which = which,
+                                  nThread = nThread,
                                   verbose = verbose)
     }
     #### Cast wider ####
@@ -129,10 +69,33 @@ r_repos_data <- function(include=NULL,
         meta_desc <- description_extract_multi(pkgs = unique(pkgs$package),
                                                nThread = nThread,
                                                verbose = verbose)
+        by <- base::intersect(names(meta_desc),names(pkgs))
         pkgs <- data.table::merge.data.table(meta_desc,
                                              pkgs,
                                              all = TRUE,
-                                             by="package")
+                                             by=by)
+    }
+    #### Add GitHub metadata ####
+    #### Get repo metadata ####
+    if(isTRUE(add_github)){
+        if(!all(c("owner","repo") %in% names(pkgs))){
+            messager("Skipping add_github, as no owner/repo data available.",
+                     v=verbose)
+        } else {
+            meta_gh <- lapply(seq_len(nrow(pkgs)), function(i){
+                tryCatch({
+                    echogithub::github_metadata(owner = pkgs[i,]$owner,
+                                                repo = pkgs[i,]$repo,
+                                                add_traffic = TRUE,
+                                                verbose = verbose)
+                }, error=function(e){messager(e,v=verbose);NULL})
+            }) |> data.table::rbindlist(fill = TRUE)
+            by <- base::intersect(names(meta_gh),names(pkgs))
+            pkgs <- data.table::merge.data.table(meta_gh,
+                                                 pkgs,
+                                                 all = TRUE,
+                                                 by=by)
+        } 
     }
     return(pkgs)
 }