+++ R/r_repos_downloads.R
@@ -2,6 +2,8 @@
 #' 
 #' Get the total number of downloads for each R package in 
 #' CRAN, Bioconductor, and/or GitHub.
+#' @source \href{https://github.com/GuangchuangYu/dlstats/issues/5}{
+#' \pkg{dlstats} Issue with getting downloads data for many packages at once.}
 #' @param pkgs Output from \link[echogithub]{r_repos_data}.
 #' @param verbose Print messages.
 #' @inheritParams r_repos
@@ -11,17 +13,30 @@
 #' @importFrom data.table rbindlist data.table setnames dcast merge.data.table
 #' @importFrom data.table :=
 #' @examples 
+#' #### All packages ####
+#' pkgs_all <- r_repos_downloads(which = "Bioc")
+#'
+#' #### Specific packages ####
 #' pkgs <- r_repos_data()[r_repo=="CRAN",][seq_len(5),]
-#' pkgs2 <- r_repos_downloads(pkgs = pkgs)
-r_repos_downloads <- function(pkgs,
+#' pkgs2 <- r_repos_downloads(pkgs = pkgs,
+#'                            which = c("CRAN","Bioc"))
+r_repos_downloads <- function(pkgs=NULL,
                               which=r_repos_opts(),
                               use_cache=TRUE,
+                              nThread=1,
                               verbose=TRUE){ 
+    # templateR:::source_all()
+    # templateR:::args2vars(r_repos_downloads)
     
-    requireNamespace("dlstats")
+    #### Avoid connection errors ####
+    # closeAllConnections() 
     downloads <- NULL;
     
     #### Check args ####
+    if(is.null(pkgs)) {
+        pkgs <- r_repos_list(which = which, 
+                             verbose = verbose)
+    }
     pkgs <- check_pkgs(pkgs = pkgs)    
     if(all(is.null(which))) return(NULL)
     which <- tolower(which)
@@ -38,6 +53,7 @@ r_repos_downloads <- function(pkgs,
     if("cran" %in% which){
         res[["CRAN"]] <- r_repos_downloads_cran(pkgs=pkgs,
                                                 use_cache=use_cache,
+                                                nThread=nThread,
                                                 verbose=verbose)
     }  
     #### Bioc ####