+++ R/r_repos_downloads_bioc.R
@@ -1,23 +1,35 @@
 r_repos_downloads_bioc <- function(pkgs,
-                                   use_cache=TRUE,
+                                   use_cache=TRUE, 
+                                   min_batch_size=10,
+                                   method=c("biocpkgtools",
+                                            "dlstats"),
+                                   nThread=1,
                                    verbose=TRUE){
+    
     r_repo <- NULL;
     
-    messager("Gathering Bioc downloads data.",v=verbose)
+    method <- tolower(method)[1]
+    messager("Gathering Bioc downloads data with",
+             paste0("method=",shQuote(method)),v=verbose)
     null_dt <- data.table::data.table(package=pkgs$package,
                                       downloads=NA)
     pkgs <- pkgs[r_repo=="Bioc",]
-    if(nrow(pkgs)==0) return(null_dt)
-    bioc <- (dlstats::bioc_stats(packages = pkgs$package,
-                                 use_cache = use_cache) |>
-                 data.table::data.table() 
-    )
-    if(nrow(bioc)>0){
-        bioc <- bioc |>
-            data.table::setnames(c("Nb_of_distinct_IPs","Nb_of_downloads"),
-                                 c("downloads_unique","downloads"),
-                                 skip_absent = TRUE)
-    } else{
+    if(nrow(pkgs)==0) return(null_dt) 
+    t1 <- Sys.time()
+    if(method=="dlstats"){
+        bioc <- r_repos_downloads_bioc_dlstats(pkgs=pkgs,
+                                               use_cache=use_cache,
+                                               nThread=nThread,
+                                               min_batch_size=min_batch_size, 
+                                               verbose=verbose)
+    } else if(method=="biocpkgtools"){
+        bioc <- r_repos_downloads_bioc_biocpkgtools(pkgs = pkgs)
+    }  
+    #### Report ####
+    report_time(start = t1, verbose = verbose)
+    #### Check ####
+    if(nrow(bioc)==0){ 
+        messager("No downloads data retrieved for any Bioc packages.",v=verbose)
         bioc <- null_dt
     }
     return(bioc)