+++ R/r_repos_downloads_github.R
@@ -2,7 +2,7 @@ r_repos_downloads_github <- function(pkgs,
                                      fields=NULL, 
                                      verbose=TRUE){
     github_url <- username <- package_name <- 
-        downloads <- clones_count <-  NULL; 
+        downloads <- clones_count <-  owner <- repo <- package <- NULL; 
     
     null_dt <- data.table::data.table(package=pkgs$package,
                                       downloads=NA)
@@ -16,16 +16,14 @@ r_repos_downloads_github <- function(pkgs,
     pkgs2 <- lapply(pkgs$package, function(p){
         if(p %in% github$package_name){ 
             messager("Gathering GitHub downloads data with",
-                     "githubinstall::gh_list_packages :",p,
-                     v=verbose)    
+                     "githubinstall:",p,v=verbose)    
             github_i <- subset(github, package_name==p)
             return(data.table::data.table(package=p,
                                           owner=github_i$username,
                                           repo=github_i$package_name))
         } else {
             messager("Gathering GitHub downloads data with",
-                     "echogithub::description_extract :",p,
-                     v=verbose)    
+                     "echogithub:",p,v=verbose)    
             tryCatch({ 
                 description_extract(repo = p,
                                     fields = fields,
@@ -49,11 +47,13 @@ r_repos_downloads_github <- function(pkgs,
     }
     #### Get GitHub traffic metadata ####
     traffic <- lapply(seq_len(nrow(pkgs2)), function(i){
-        tr <- github_traffic(owner = pkgs2[i,]$owner,
-                             repo = pkgs2[i,]$repo,
-                             verbose = verbose)
-        if(!is.null(tr)) tr[,package:=(pkgs2[i,]$package)]
-        return(tr)
+        tryCatch({
+            tr <- github_traffic(owner = pkgs2[i,]$owner,
+                                 repo = pkgs2[i,]$repo,
+                                 verbose = verbose)
+            if(!is.null(tr)) tr[,package:=(pkgs2[i,]$package)]
+            return(tr)
+        }, error=function(e){messager(e,v=verbose); NULL})
     }) |> data.table::rbindlist(fill=TRUE)
     #### Check for empty data ####
     if(nrow(traffic)>0) {