+++ R/zzz.R
@@ -1,3 +1,5 @@
 .onLoad <- function(libname, pkgname){
+    requireNamespace("data.table")
     .datatable.aware <- TRUE 
+    data.table::setDTthreads(threads = 1)
 }